fixed sword issue
lots of issues coming from player's scale being set to 0.1 rather than scaling down the sprite.  I added in some counter measures incase we decide to keep it scaled to 0.1, which will fix our sword issue, but just a heads up that is where the issue came from as we begin to add more weapons in.  I also made it so the temp sword animation isnt hugging directly on the players body.